Add historic induction among other small changes

+++ src/recon/reconciliation.py
@@ -1,6 +1,10 @@
 import filecmp, json, os
+import hashlib, uuid, shutil
+from PIL import Image
 from src.dao import create
 from src.dao.image import ForumImage
+from src.dao.thread import ForumThread
+from sqlitedao import DuplicateError
 
 
 with open('conf/conf.json') as conf:
@@ -8,10 +12,12 @@ with open('conf/conf.json') as conf:
 
 
 # Main func
-def recon_id(author_id):
+def recon_id(author_id, remove_duplicate=False):
+    if remove_duplicate:
+        remove_duplicate=True
     dao = create.get_dao()
     images = dao.get_items(ForumImage, {"author_id": author_id})
-    crawled_images = [im for im in images if im.get_crawled_status()]
+    crawled_images = [im for im in images if im.get_crawled_status() and not im.is_duplicate()]
     print("Found {} images to recon".format(len(crawled_images)))
     failed_images = []
     for image in crawled_images:
@@ -19,6 +25,7 @@ def recon_id(author_id):
         if not os.path.isfile(real_path):
             print("{} is missing".format(real_path))
             failed_images.append(image)
+            continue
         if filecmp.cmp(FAIL_FILE_PATH, real_path):
             print("{} has failed".format(real_path))
             os.remove(real_path)
@@ -28,6 +35,7 @@ def recon_id(author_id):
     print("Found {} images in wrong state".format(len(failed_images)))
     if failed_images:
         dao.update_items(failed_images)
+    reconcile_duplicates(author_id, remove_duplicate)
 
 
 # Main func
@@ -38,6 +46,8 @@ def remove_id_images(author_id):
     print("Found {} images to remove".format(len(crawled_images)))
     pardirs = set()
     for image in crawled_images:
+        if image.is_duplicate():
+            continue
         real_path = image.get_file_path()
         pardirs.add(os.path.dirname(real_path))
         os.remove(real_path)
@@ -63,4 +73,107 @@ def reconcile_multiple_auth_name(limit=500):
                 im_set[image.get_author()] = 1
         if len(im_set) > 1:
             print("Differences in auth_names identified: {}".format(im_set))
-    
+
+
+def hashstr(fname):
+    hash_md5 = hashlib.md5()
+    with open(fname, "rb") as f:
+        for chunk in iter(lambda: f.read(8192), b""):
+            hash_md5.update(chunk)
+    return hash_md5.hexdigest()
+
+
+def reconcile_duplicates(author_id, remove=False):
+    dao = create.get_dao()
+    images = dao.get_items(ForumImage, {"author_id": author_id})
+    crawled_images = [im for im in images if im.get_crawled_status() and not im.is_duplicate()]
+    print("Found {} images to recon".format(len(crawled_images)))
+    hash_records = {}
+    for image in crawled_images:
+        real_path = image.get_file_path()
+        hastr = hashstr(real_path)
+        if hastr in hash_records:
+            hash_records[hastr].append(image)
+        else:
+            hash_records[hastr] = [image]
+    for hastr, imlist in hash_records.items():
+        if len(imlist) > 1:
+            print("Found duplicate imgs: {}".format(hastr))
+            for image in imlist:
+                print("{} belongs to {}".format(image.get_file_path(), image.get_href()))
+            if remove:
+                kept = imlist.pop()
+                print("Keeping: {}".format(kept.get_file_path()))
+                for dup in imlist:
+                    os.remove(dup.get_file_path())
+                    dup.mark_as_duplicate()
+                dao.update_items(imlist)
+
+
+def add_historic_thread(dao, author, author_id, flist):
+    href = "/h/" + author_id
+    title = author + "_history"
+    history_thread = ForumThread(href=href, replies=365, title=title, author=author, author_id=author_id, create_time="1991-12")
+    history_thread.mark_as_crawled(len(flist))
+    try:
+        dao.insert_item(history_thread)
+    except DuplicateError:
+        print("Historic thread already exists")
+    add_historic_images(dao, flist, href, author, author_id)
+
+
+def add_historic_images(dao, flist, href, author, author_id):
+    imlist = []
+    for f in flist:
+        try:
+            src = "history://" + f.split("/")[-1]
+            image = Image.open(f)
+            width, height = image.size
+            size = os.stat(f).st_size
+            imageria = ForumImage(src=src, href=href, image_size=size, image_width=width, image_height=height, author=author, author_id=author_id, floor_num=69, create_time="1991-12")
+            imageria.mark_as_crawled()
+            imageria.set_uuid(str(uuid.uuid4()))
+            new_file_path = imageria.get_file_path()
+            print("Moving {} to {}".format(f, new_file_path))
+            shutil.move(f, new_file_path)
+            imlist.append(imageria)
+        except Exception as e:
+            print("Operation for {} failed due to {}".format(f, e))
+    if imlist:
+        print("Inserting {} converted images".format(len(imlist)))
+        dao.insert_items(imlist)
+
+
+# Main func
+def induct_history(author_id):
+    dao = create.get_dao()
+    # Find all inductee
+    images = dao.get_items(ForumImage, {"author_id": author_id})
+    # Move forward if all author agree:
+    author_set = set()
+    for image in images:
+        author_set.add(image.get_author())
+    if len(author_set) > 1:
+        print("Conflicting author information, exiting")
+        return
+    author = author_set.pop()
+    crawled = [im for im in images if im.get_crawled_status() and not im.is_duplicate()]
+    uuid_set = {im.get_uuid() for im in crawled}
+    if not crawled:
+        print("No crawled images belong to this author")
+        return
+    ops_dir = os.path.dirname(crawled[0].get_file_path())
+    img_files = [f for f in os.listdir(ops_dir) if f.endswith("jpg")]
+    historics = list(filter(lambda f: f.split(".jpg")[0] not in uuid_set, img_files))
+    print("Found {} historic files to induct".format(len(historics)))
+    if not historics:
+        return
+    for f in historics:
+        fname = os.path.join(ops_dir, f)
+        print("FILEPATH: {}".format(fname))
+    if not str(input("Convert? y/n")) == "y":
+        print("Not converting history today")
+        return
+    flist = [os.path.join(ops_dir, f) for f in historics]
+    add_historic_thread(dao, author, author_id, flist)
+
